faulthandler: extract into module/faulthandler/ following the _socket pattern
importing.rs shrinks by 307 lines.
