+++ pyre/pyre-interpreter/src/module/faulthandler/mod.rs
@@ -1,0 +1,8 @@
+//! faulthandler module — PyPy: pypy/module/faulthandler/
+//!
+//! Pyre has no Python-level traceback machinery yet, so signal handlers
+//! write a short "Fatal Python error: <name>" line to fd 2 and restore the
+//! default disposition + reraise the signal so the process dies normally.
+
+pub mod interp_faulthandler;
+pub mod moduledef;
